// opensilex-front/front/src/lib/model/frontConfigDTO.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FrontPluginConfigDTO } from './frontPluginConfigDTO';


export interface FrontConfigDTO { 
    plugins?: { [key: string]: FrontPluginConfigDTO; };
    welcomeComponent?: string;
    homeComponent?: string;
    notFoundComponent?: string;
    headerComponent?: string;
    loginComponent?: string;
    menuComponent?: string;
    footerComponent?: string;
    menu?: { [key: string]: string; };
}